backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
pydantic[email]==2.5.2
fastf1==3.6.0
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
emails==0.6